Add icudt.dll for Windows
1. Generate and add windows/icudt.dll with the procedure outlined
in README.chromium. It uses a out-of-tree copy of the upstream ICU
along with our custom-built icudtl.dat and a locally modified version
of makedata.mak.

We used to have a separate build/ directory for VS solution/project files
to build icudtl.dll. Maintaining them is rather cumbersom now that we
want to update our ICU (major version changes) more frequently. 

Note that icudt.dll is not used by default (icu_use_data_file_flag=1). 
The GN build still uses it by default and we should not break that build.

2. Add scripts/make_mac_assembly.sh to simplify the generation of the icu
data assembly source file for Mac.

3. Update README.chromium accordingly.

This CL was uploaded and reviewed at 

https://codereview.chromium.org/255943004/

Due to a malfunction at codereview.chromium.org, I'm landing this CL 
manually.

BUG=132145
TEST=None until icu is rolled to this version.


git-svn-id: https://src.chromium.org/chrome/trunk/deps/third_party/icu52@266601 4ff67af0-8c30-449e-8e8b-ad334ec8d88c