Add back display names for non-UI languages in A-L list
I was too aggressive in trimming the data and dropped the display
names for languages that Chromium needs (for non-UI languages
that are in the A-L list). It's not my intention (the comment in
trim_data.sh said one thing, but the code did another). 

Besides, add Norweigian (nb) and Malay (ms) locale data that were not 
included by mistake.

Also update trim_data.sh script NOT to drop 'ALIAS' lines which are
used to indicate that a given locale is an alias to another locale.
That also required adding ro_MD.txt (null locale which mo.txt is 
aliased to).

The above three adds about 110kB to the icu data (from 10.3MB to 10.4MB).

Also update the pre-built icu data files for Linux, Mac and Windows.
The Android data will be updated in a follow-up patch.

BUG=132145
TEST=When ICU is rolled, unit_tests:ExtensionL10* pass.
TBR=mark

Review URL: https://codereview.chromium.org/264973016

git-svn-id: https://src.chromium.org/chrome/trunk/deps/third_party/icu52@268285 4ff67af0-8c30-449e-8e8b-ad334ec8d88c