Remove invalid link_settings from target condition.
ICU fails to gyp when run with GYP_DEFINES="android_webview_build=1
use_system_icu=0 use_system_stlport=1" which is a combination of
settings we're trying to bring up temporarily as we migrate away from
system libraries. It fails because it's not permitted to specify
link_settings in a target_condition as the processing is too late.

Remove the invalid link_settings, since we can't do this outside
target_conditions as it's not valid to use -lgabi++ on the host build of
ICU. The link dependency on gabi++ will have to be satisfied manually in
the main libwebviewchromium target instead for this configuration.

BUG=409851
R=mkosiba@chromium.org
TBR=jshin@chromium.org

Review URL: https://codereview.chromium.org/527193003

git-svn-id: https://src.chromium.org/chrome/trunk/deps/third_party/icu52@291781 4ff67af0-8c30-449e-8e8b-ad334ec8d88c